[updates] Migrate `IUpdatesController` to use coroutines (#35809)
# Why
Currently, we rely a lot on callbacks in updates. Reducing them will be a multi step process. This PR starts with the `IUpdatesController` interface.

# How
Convert the functions related to checking and fetching updates to suspend functions. I've also removed the use of `wait` and `notify` in favour of `CompletableDeferred`. Note, there is an upstream issue that is currently causing a failure in `ErrorRecovery.kt`. I've addressed it [here](https://github.com/facebook/react-native/pull/50400) but it won't be in RC4.

# Test Plan
Running bare-expo in release mode and CI