[android][updates] Refactor `DatabaseHolder` (#36004)
# Why
Refactors the `DatabaseHolder` to make it's usage more consistent and idiomatic to kotlin. The intention of this class is to be a mechanism for access control to the database. The drawbacks were it used `wait` and `notify` and I noticed in the tests that it can cause deadlocks. 

# How
Used a mutex instead of `wait` and `notify`. I've also removed `releaseDatabase`. This made usage error prone and easy for someone to forget. `mutex.withLock` will release automatically once the database operation is completed. 

# Test Plan
Running in release working as expected, e2e tests are passing locally, CI. This fixes the tests in the downstack PRs