[ios][background-task][expo-go] reverted and fixed background task
A previous PR #35969 was merged that removed the `processing` key from the `UIBackgroundModes` in Expo Go's Info.list file.

There is no real need for removing background tasks from Expo Go since it will work (tested and verified). Expo background fetch was also available in Expo Go.

This was not necessary, since the error we got from AppStore was that we were missing another key when `processing` was set - the `BGTaskSchedulerPermittedIdentifiers` key.

This commit fixes this by reverting the change to the `et eas remove-background-permissions-from-info-plist` script, and adding the missing `BGTaskSchedulerPermittedIdentifiers` so that `expo-background-task` can run in Expo Go.