[docs][task-manager] fix docs on Expo Go usage
Our documentation for the Task Manager contains some unclear instructions about usage in Expo Go that isn't correct. The Task Manager docs shouldn't contain any limitations for Expo Go usage - this is the responsibility of the consuming libraries.

This commit fixes this by adding a note about Expo Go and reference to the libraries that uses this package.

An issue was also mentioning this: #33776