[ios][epxo-go] removed local network access test on simulator
After we've upgraded to latest Xcode 16.3 and MacOS Sequioa 5.4 (not sure which since it is a bit hard to roll back and test) we have gotten a lot of feedback saying that Expo Go doesn't allow opening and connecting to apps when it runs on the Simulator.

It complains with the dialog "Local network access required" which indicates that it couldn't display the local network access permissions dialog.

This works fine when running on a device, but not in the Simulator.

This commit fixes this by avoiding the permission check when code is complied for the simulator.

Apple has a document that contains a note about local network privacy not being supported by the Simulator here:

https://developer.apple.com/documentation/technotes/tn3179-understanding-local-network-privacy

And there has been similar messages in the Apple dev forums:

https://developer.apple.com/forums/thread/766133

(look at the last message from "peterz")

Closes ENG-15446