fix(cli): fix frame showing and react canary  (#36053)
This pull request includes several changes to the `packages/@expo/cli`
to fix bugs, improve error messages, and enhance functionality. The most
important changes include ensuring the fast resolver is enabled when
React Canary is enabled, fixing the stack trace formatting, and updating
the changelog with the latest changes.

### Bug Fixes and Enhancements:

*
[`packages/@expo/cli/src/start/server/metro/instantiateMetro.ts`](diffhunk://#diff-362d97681d5ead3b0b4b966473c0b148267e7dc68dba8af9734dfdfe3e0ad05dR87-R97):
Ensured the fast resolver is enabled when React Canary is enabled by
adding a check for `isReactCanaryEnabled` and setting the
`EXPO_USE_FAST_RESOLVER` environment variable accordingly.
[[1]](diffhunk://#diff-362d97681d5ead3b0b4b966473c0b148267e7dc68dba8af9734dfdfe3e0ad05dR87-R97)
[[2]](diffhunk://#diff-362d97681d5ead3b0b4b966473c0b148267e7dc68dba8af9734dfdfe3e0ad05dL153-R164)

*
[`packages/@expo/cli/src/start/server/metro/metroErrorInterface.ts`](diffhunk://#diff-1b519a55a14b36c361c0683bcfa674fe8b8bfa190d7739ac78bb71ccad7af10cL150-R150):
Fixed the stack trace formatting issue by correcting the condition to
show collapsed frames.

### Documentation Updates:

*
[`packages/@expo/cli/CHANGELOG.md`](diffhunk://#diff-dced9f7f9eaad1257a4d1786ee6facb5010790cec922eb8662e33a83923c7f7dR11-R12):
Added an entry to ensure the fast resolver is enabled when React Canary
is enabled.

*
[`packages/@expo/cli/CHANGELOG.md`](diffhunk://#diff-dced9f7f9eaad1257a4d1786ee6facb5010790cec922eb8662e33a83923c7f7dL28-R30):
Corrected the formatting of the changelog entry for the helpful
recommendation added to the "Xcode not installed" error message.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>