[expo][2/n] Roll out edge-to-edge: Add `react-native-edge-to-edge` as a dependency of `expo` (#35812)
# Why

We need all projects to support configuring the edge to edge feature in
SDK53. We use `react-native-edge-to-edge` for that, so we add is as a
dependency of `expo`.
It is not an expo module so it won't be autolinked as it is a transitive
dependency for the project. I have added a special case to autolinking,
where we check if the `expo` version is greater than `52` and if it is
`react-native-edge-to-edge` is added to the list of autolinked packages.

# How

Added `react-native-edge-to-edge` as a dependency of the `expo` package
Added a special case in autolinking that checks if the expo version is
high enough and adds the `react-native-edge-to-edge` to autolinking.

# Test Plan

Tested in a seperate project, which had `edge-to-edge` added as a
transitive dependency via a random 3rd party package, and patched
autolinking to contain code from this PR (with a change of expo version
from 53 to 52 so that the `if` gets triggered). Edge-to-edge has
installed and autolinked successfuly.
Also worked for BareExpo and Expo Go.