[expo][3/n] Roll out edge-to-edge: Add `enableEdgeToEdge` app config field (#35958)
# Why

We want users to be able to configure `edge-to-edge` from the app config
file.

# How

We use `react-native-edge-to-edge` for enabling `edge-to-edge`.
If `enableEdgeToEdge` is false or undefined
- We warn about a future drop of support
- Disable the force-enable for Android 15
If `enableEdgeToEdge` is true
- Import `react-native-edge-to-edge` config plugin and apply it with
some reasonable defaults (DayNight theme)
If someone configured `react-native-edge-to-edge` plugin directly
- Apply the plugin as normal

# Test Plan

Tested in a test project created in `expo/apps` 
I also tested interaction with `androidStatusBar` properties, and they
work as expected (`android:statusBarColor`) doesn't work in edge-to-edge
mode.