[expo][4/n] Roll out edge-to-edge: Enable by default in new projects (#35959)
# Why

Since edge-to-edge is always in enforced in Android 16+ we want to
enable edgeToEdge by default in new projects to have as many people
onboard as possible.

# How

Updated the template files.

**should we also update the minimal template?**
Please let me know if there is anything else that has to be updated to
change the templates.

# Test Plan

Ran `npx create-expo-app -t <local-path-to-template-directory>`
I'm not sure if there is a better way of testing this