[expo-router][next] Bump expo-router @radix-ui/react-slot dependency
The @radix-ui/react-slot@1.1.0 is the first version to support react@19.
Currently the expo-router depends on @radix-ui/react-slot@1.0.1 that
does not list react@19 as a peer dependency, thus making npm install
both react@18 and react@19 in the project, causing errors unless
overriding react in main package.json.

The changes from @radix-ui/react-slot@1.0.1 to
@radix-ui/react-slot@1.1.0 are mostly in the way to access component
refs, the full diff can be reviewed here:
https://github.com/radix-ui/primitives/compare/6343181bd1fb9c0962bb1190dadd1e5e10b32fd4...55f9189a4b3e1a525c89297d5f337c03b2a44f61?w=#diff-444aece5662ae3d455daa84d8be1dafa6c878ca67278a99556c830b7f461b155