[status-bar] Use react-native-edge-to-edge system bars when edge-to-edge is enabled (#36087)
# Why

To make the migration to edge-to-edge easier

# How

There is actually not a lot going on here, it just looks like a big
change because I had to shuffle the files around.

-
[types.ts](https://github.com/expo/expo/blob/brent/eng-15467-make-expo-status-bar-use-systembars-when-edge-to-edge-is/packages/expo-status-bar/src/types.ts)
and
[NativeStatusBarWrapper.tsx](https://github.com/expo/expo/blob/brent/eng-15467-make-expo-status-bar-use-systembars-when-edge-to-edge-is/packages/expo-status-bar/src/NativeStatusBarWrapper.tsx)
- I split these out from the former main StatusBar component file. I
wanted to be able to isolate any Android specific logic around
edge-to-edge to an `.android.tsx` file, but still re-use the normal
implementation when edge-to-edge is not enabled.
-
[StatusBar.android.tsx](https://github.com/expo/expo/blob/brent/eng-15467-make-expo-status-bar-use-systembars-when-edge-to-edge-is/packages/expo-status-bar/src/StatusBar.android.tsx)
- this imports `NativeStatusBarWrapper` and wraps each of its methods,
so if edge-to-edge is enabled it just calls those directly, otherwise it
now has different behavior.
- Tests had to be re-organized a bit.
- Otherwise everything is basically the same.