[ios][expo-go] add logging of manifest exceptions
When manifest exceptions happens in Expo Go we aren't starting/loading the bundle in a JS context, meaning we don't log errors to the packager - just displaying the ErrorView.

This commit fixes this by introducing a helper class that can log to the packager without a JS context in Expo Go.