[video][android] Fix touch events not being emitted for `VideoView` (#35479)
# Why

Fixes https://github.com/expo/expo/issues/34630
The `PlayerView` from `ExoPlayer` seems to be handling the touch events
in a way that prevents them from being dispatched to RN.

# How

Created a workaround with `onTouchEvent` and `onInterceptTouchEvent` in
order to re-dispatch correct events to the RN side when they are
received.
I had to rename `VideoView.id` to `VideoView.videoViewId`, because I was
unknowingly overriding an existing id used by React Native for view
identification.

# Test Plan

Tested in BareExpo on an Android 15 device