[media-library][2/2] Make moving assets unnecessary - Add `initialAssetLocalUri` parameter to `createAlbumAsync`  (#35692)
# Why

In part [1/2] we made it possible to pass the album as a parameter for
the asset. In case of a fresh project there is still an issue of not
having the album created, and not having any assets to put into the
`assets` parameter of `createAlbumAsync` . This creates chicken and egg
problem, because we also don't have an album to put in the `album`
parameter of `createAlbumAsync`.

The only option that allows us to skip moving assets unnecessarly is to
make `createAlbumAsync` accept a `uri` of an initial file that is meant
to be stored in the album.

# How

Added `initialAssetLocalUri` parameter to `createAlbumAsync`.
Also added two new tests that test new the functionalities updated the
docs and added a note about moving the assets.

# Test Plan

All MediaLibrary tests pass on Android and iOS