[navigation-bar] Add support for methods from `react-native-edge-to-edge` (#36163)
# Why

Our existing methods are partially or fully incompatible with
`react-native-edge-to-edge`

# How

- Added `react-native-edge-to-edge` as a dependency. Used it in
`setVisibilityAsync` when edge-to-edge is enabled, otherwise fall back
to the default implementation.
- Moved existing methods into a separate file so that there is a
distinction between the edge-to-edge stuff
- Added the `setStyle` method only for edge-to-edge, since it is
distinctive from anything we currently have in the module (there is
`setButtonStyleAsync` and `setBackgroundColorAsync`, but those methods
are different. (Maybe we could stretch the button style 🤔 )

# Test Plan

Tested in a new app created in the monorepo on Android and iOS