[docs] Updates custom initialization changes for SDK 53 (#36216)
# Why

The existing doc is correct for SDK 52, but needs to be updated for SDK 53, as iOS has moved to using the new `ReactNativeFactory` in RN 0.79.

# How

Added SDK 52 and SDK 53 tabs in the doc, with the correct sample code for each for iOS. No changes for Android.

# Test Plan

Docs CI should pass

# Checklist


- [x] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)