chore(cli): Replace `internal-ip` with `lan-network` (#35102)
# Why

We've previously added a fallback for `internal-ip` in #34043 after
discovering that it fails on Windows when some utilities aren't
installed, and pinned it to a new specific version in #34325. However,
it would still be nice to replace it with a package that is
cross-platform compatible (supports all platforms that Node.js supports)
without executing external system programs and utilities. Additionally,
it'd be great to remove its dependency chain.

We'd like to switch to the fallback implementation fully, but the
fallback implementation fails if there are no known routes on the
machine (if it's offline, or if the network is offline / has no route to
the public internet)

# How

I've pulled out the fallback implementation into a separate package,
[`lan-network`](https://github.com/kitten/lan-network). This package
contains two additional ways to determine the LAN IPv4 assignment.

It first uses our previous approach. It creates a dgram UDP4 socket and
connects to a publicly routable IP on the internet. Since this is UDP
and no packets are sent no actual connection is established. But the
socket is then queried for its local address, which will always be the
address that's associated with the "default gateway route" to the
internet. That's then matched up with the available IPv4 network
interfaces.

If the prior fails, we don't have a default gateway, but we can still
guess what the default gateway would have been using DHCP, since most
routers and LANs have a DHCP server. This is done by broadcasting a
`DHCPDISCOVER` packet scoped to the known addresses, while listening for
replies. If a reply is received, the address of the responder is matched
to the IPv4 network interfaces, and we assume that that's the default
gateway.

Lastly, if the prior two methods fail, we use a sorted list of network
interface assignments. External interfaces are prioritised. After, known
internal IP ranges are prioritised. We simply pick the first assignment
from this list, since it's likely to either be:
- a local network IP (or at least usable locally), or a VM IP
- an interface that's "larger" than the `127.0.0.1` local IP, e.g. a VPN
- or just the local network IP (e.g. `127.0.0.1`)

# Test Plan

- [x] Test on macOS
  - [x] Disconnected and offline
  - [x] Connected and offline
  - [x] Connected to a VPN
  - [x] Connected to a VPN as default gateway
- [x] Test on Windows
  - [x] Disconnected and offline
  - [x] Connected and offline
  - [ ] Connected to a VPN
  - [ ] Connected to a VPN as default gateway
- [x] Test on Linux
  - [x] Disconnected and offline
  - [x] Connected and offline
  - [x] Connected to a VPN
  - [x] Connected to a VPN as default gateway

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)