[expo-go][ios] add logging of manifest exception (#36170)
# Why

When manifest exceptions happens in Expo Go we aren't starting/loading
the bundle in a JS context, meaning we don't log the errors to the
packager - they're caught and displayed in the ErrorView:

<img width="300" alt="image"
src="https://github.com/user-attachments/assets/61cd7716-1bf2-496b-b099-6f01f4098523"
/>

Resolves ENG-15442 together with #36171

# How

This commit fixes this by introducing a helper class that can log to the
packager without a JS context in Expo Go:

- Added `Exponent/Kernel/DevSupport/EXPackagerLogHelper.m/h` for
connecting and logging to the packager. It logs a single message using a
static method and then closes the web socket.
- Added call to log the error inside `Exponent/Kernel/Views/ErrorView.m`
when the error-prop is updated

<img width="861" alt="image"
src="https://github.com/user-attachments/assets/c97ba4fe-6a1b-4a0c-bf58-95d152621637"
/>

# Test Plan

- Run Expo Go and try to open a running app with SDK-52 against `main`
- Observe that Expo Go displays and error and that the same error is
displayed in the CLI
 
# Checklist

- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).