Add missing module: "preserve"
We don't use tsc to modify files, but if it's used, we must preserve
the module imports/exports.

If `target` isn't set it defaults to `es5`, which makes `module`
to `CommonJS`. However, we don't set the target, but want to allow
ES module syntax. This means `preserve` is the best option here,
which is more equivalent to `target: es2015`