Filter out non-errors and TS6046 in evaluateTsConfig
We shouldn't throw an error in `evaluateTsConfig` if the diagnostic
category is not `1` (DiagnosticCategory.Error = 1).

We also shouldn't throw on TS6046. When an outdated TypeScript version
is used (for example 5.2.2) some `tsconfig.json` compilerOptions are
invalid. Our preset now contains `module: 'preserve'` which doesn't
exist in older versions of TypeScript.

This will be also thrown when `tsc` is run, however, we can stop it
from breaking `expo export`, `expo export:embed`, and `expo start`, by
ignoring it. This is safe to ignore since we only use `paths` and
`baseUrl` from evaluated configs.