[expo-router] Replace "shadow*" style props with "boxShadow" style prop
Currently, when using the expo-router@^5 and one of the included
layouts, this warning is shown when navigation to any screen: `"shadow*"
style props are deprecated. Use "boxShadow".`

The only place I could found the "shadow*" props use is on Sitemap.tsx,
and indeed replacing them with "boxShadow" gets rid of the warning.