[video] Add `contentType` field to `VideoSource` (#36234)
# Why

Fixes https://github.com/expo/expo/issues/35823

Android ExoPlayer determines the video type only based on the uri. When
playing HLS video, which doesn't have `m3u8` or `Manifest` in the uri
the playback will fail, because it's unable to determine the content
type.

# How

Added a field to `VideoSource`, where the users can specify what is the
type of the video.
This is also helpful on iOS where previously we were checking if uri
contains `m3u8` to determine whether to fetch video tracks. I also added
some warnings to the docs.

# Test Plan

Tested in BareExpo on Android and iOS.