[notifications] support android action buttons when not in foreground (#35295)
# Why

Historically, when an Android app was terminated and there was a notification with action buttons (as defined by `category`), tapping the button would be a noop. This PR changes that - users can use [`registerTaskAsync`](https://docs.expo.dev/versions/latest/sdk/notifications/#registertaskasynctaskname) and _on Android only_ the task will be executed. On iOS, users can use `NotificationResponseReceivedListener` already (in all app states) but this approach doesn't work for Android, which is why I opted for using the background task. This could be unexpected so I marked it as a breaking change. I'm happy to accept other / better ideas :).

# How

- change the android native impl to run tasks also in response to action button interaction
- remap the task payloads to align better across platforms. They are still quite different but there's more common ground than before. Also added TS type for the task payload shape.

# Test Plan

- notification tester app
- see the task at https://github.com/expo/expo/blob/b229787edb83306d3fdb94764e7ff289194917cf/apps/notification-tester/src/registerTaskAsync.ts#L67 - it runs when notification is received and also when action button is pressed. I used the sync methods from `expo-sqlite/kv-store` to store the event payloads in all application states and later retrieved them to verify the task was executed.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)