[notifications] correctly include NotificationContent.data in NotificationHandler (#36195)
# Why

fixes a bug where `NotificationContent.data` was not being correctly included in the `NotificationHandler` events. The notification data passed to the handler wasn't being added because `mapNotificationContent` adds `data` by JSON parsing `dataString`.

# How

Modified the `NotificationsHandler.ts` file to use the `mapNotification` utility function when passing notifications to the handler, same as in `addNotificationReceivedListener`. 


# Test Plan

- notification tester

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)