[expo-go][ios] update reversed links in error texts (#36304)
# Why

Markdown links in error text was formatted in the opposite way (url
first, name last), this worked when displaying in iOS since the parser
was implemented to parse using the wrong order.

# How

This commit fixes this by changing the URLs and the parser.

# Test Plan

- Run Expo Go and verify that links are displayed correctly in the Error
View (try loading an SDK-52 bundle using the current main Expo Go)

# Checklist

- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).