[docs] Add missing `extends` properties in `tsconfig.json` samples (#36306)
# Why

We're stating:
> A project's **tsconfig.json** should extend `expo/tsconfig.base` by
default

But then go on to show `tsconfig.json` examples that don't contain this
property. If a user isn't reading carefully and copies the examples,
they'll be missing this crucial addition. The preset contains everything
a user needs to add to make a config valid, so by simply adding this to
the samples, all the samples are immediately copyable.

## Side-note

I've noticed this section below the changes:
https://docs.expo.dev/guides/typescript/#typescript-for-projects-config-files

I think we should discourage this and potentially alter our
`expo/tsconfig.base`, since it currently excludes these files. If a user
wishes to type check these files, it's much less error prone for them to
use `checkJs: true` or `// @ts-check` comments.

Changing configuration to TypeScript files could introduce a lot of
testing burden for us, and I'm not sure if we ever check that this setup
is valid and works across SDK versions. `ts-node` has a lot of gotchas,
and we're basically telling a user here to transpile their configs with
`ts-node` with the `tsconfig.json` that's meant for their app, which can
lead to problems.

Unrelated to this PR, but worth noting here, while I'm editing this file
:v:

# How

- Add missing `extends: 'expo/tsconfig.base` to all `tsconfig.json`
examples on TypeScript page