[background-task] remove throwing errors when calling register/unregister with already reg/unreg tasks (#36393)
# Why

Instead of throwing an Error, calling un/registerTaskAsync will now just
return if the task is already un/registered. This was introduced in
#35734

# How

Instead of throwing an error, we'll just return. The previous version
threw an exception which is not necessary.

# Test Plan

Test in BareExpo

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>