[video][android] Add an option to choose the surface type (#36212)
# Why

Fixes
https://github.com/expo/expo/issues/33789
https://github.com/expo/expo/issues/31722

When two `VideoViews` with `contentFit="cover'` are overlapping one of
the video views will 'leak' onto the surface of another. This is a
[known upstream issue](https://github.com/androidx/media/issues/1107).
Currently the only fix is to use a texture view rendering surface
instead of surface view, but that reduces the performance, so we would
like users to have a chooice.

# How

Exported two views from the `expo-video` module - one with a TextureView
surface and another with `SurfaceView` surface. On the JS side a prop
selects an appropriate view to render. The type of surface has to be
known at the time of creation of the view, so we couldn't just create a
regular prop.

# Test Plan

Tested in BareExpo on Android and iOS

<img width="326" alt="image"
src="https://github.com/user-attachments/assets/8feefdb4-7e5b-4f08-a8e9-8bc1b1bd62b6"
/>