[router] Refactor route state to utilize useStateForPath() (#36199)
# Why

React Navigation recently added the `useStateForPath()` hook that provides the minimal state needed to reconstruct a path for the current screen. This new hook is different to React Navigation's previous `getState()` functionality, as it provides the entire minimal state of all navigators, while `getState()` only provided the current and children state (not any parent state). Because of this, we were required to use the root state - which isn't always correct. The biggest problem with root state is that navigators are dynamic. Just because we tell them "this is the state" does mean they will render that state. For example with protected screens, if the state says we should render that screen but its protected the navigator will not. In React Navigation this is known as stale state and its the source of some outlying Expo Router bugs. Because of stale state, the only components that know what the real state is are the leaf screens - not the root navigator. But until now we couldn't access the root state from a leaf screen, hence why this hook is important.

The main goal of this PR is to stop using the root navigator state and instead of the `FocusedState` (the name for the minimal state) of the currently focused screen.

## Router Store

While making these changes, I also refactored the RouterStore. The RouterStore was in a bad shape and as things evolved over time it relieved heavily on `.bind(store)` for functions. I think binding functions is an anti-pattern, so it was refactored to largely remove this. We also reduced our reliance of `useSyncExternalStore`.

## useNavigation

The `useNavigation()` hook needs to access the current state of the current screen. It used to use `getStateFromPath` but can now use `useStateForPath()` to more easily access this state.

## routeInfo / getPathFromState
 
The`RouterInfo` is the object that holds the return values of `usePathname()` / `useSegments()` / etc. It currently uses  `getPathFromState`, but it needs to call this function twice and doesn't correctly handle the `FocusedState` 

Instead of rewriting this function to handle  `FocusedState` (I don't want to fork to much further), I created a new function that can parse both `NavigationState` and `FocusedState` in one pass and is much simpler, as it doesn't perform any config validation.

In the future, I may refactor `getPathFromState` to utilise this function as well, but I don't want to change too much in this PR.