fix(rsc): remove absolute paths from RSC (#35276)
# Why

- Prevent leaking absolute path data to public exports.

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- I tried moving all the logic into Metro to have a proper solution but
it depends on the ability to transform the same file multiple times with
different settings in order to create a giant graph with dependencies
that represent both server and client simultaneously. This is currently
not possible in Metro and also blocks module specifiers.
- Instead I just used paths relative to the project root and updated the
fragile system of replicating the paths during production exports.


# Test Plan

- The E2E tests should cover all existing behavior. Also ran locally on
web.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>