fix: Override transform options with `enableBabelRuntime: false` for prepend scripts (#36463)
## Why

For prepended scripts, we're transforming code that may not use any
additional dependencies, as they're initialized before the dependency
resolution part of the output bundle. However, `@babel/runtime` imports
are added via transformations which break the prepended scripts
(specifically, in JSC for `virtual/streams.js` currently)

## How

To prevent this, we should force `enableBabelRuntime` to `false` for
scripts.
This is now done in two places.

1. In `@expo/metro-config` we pass on `BabelTransformOptions.type` to
the Expo Babel caller and use it in `babel-preset-expo` to force
`enableBabelRuntime` to `false
2. For completeness (shouldn't be relevant for this case), we force
`enableBabelRuntime: false` for scripts in `metro-transform-worker`

> [!NOTE]
> This PR doesn't yet rebuild `expo-dev-launcher`, but rebundling it
shows how this change works

## Test Suite

- **Note:** I don't really know how to test this yet. The Babel preset
doesn't seem to extract the Babel runtime helpers even without this
change