fix(cli): Fix fallback module resolution strategy for non-workspace projects (#36181)
# Why

While testing, I noticed that the fallback module resolver no longer
works in non-monorepo installations (e.g. isolated dependencies with
`pnpm`). This is due to the resolver not falling back to the project
root (and upwards) any longer.

We have to provide a valid path to `originModulePath` that points to the
project root (plus a random file name since `path.dirname` is invoked on
this path)

# How

- Replace `originModulePath` with path based on `projectRoot` when
resolving `expo` and `expo-router` in fallback module resolver

# Test Plan

- Tested manually
- Not added to integration tests since workspace test works as expected,
but the non-workspace case is not officially supported yet