[docs][sdk-53] Add docs section for ES Modules Resolution (`package.json:exports`) to `metro.config.js` page (#36458)
This is an initial attempt to document `package.json:exports` support
and to provide instructions on how to disable it.

The main gist of it is:
- mention that ES Modules resolution is different
- describe how Metro uses it and falls back on old resolution strategies
- explain that `package.json:exports` may be used and what it is
- document how to disable this

We're still lacking instructions however for `import.meta.*`. It'd be
great to document it, since we don't have full support for it yet (and
namely, `import.meta.env` inline replacement), as far as I remember. But
I'm unsure how to document this and what to describe as workarounds or
common failure modes for it. (cc @Kudo @EvanBacon)

---------

Co-authored-by: Brent Vatne <brentvatne@gmail.com>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>