[router]: Fix incorrect route info for nested tabs when navigating via touch (#36558)
# Why

Fix #36549

The routeInfo is incorrectly updating after touch events for nested Tab
navigators when jumping to a screen in history

In SDK53 we changed how `routeInfo` was calculated to be on the focused
state. We then notify screens that the routeInfo has been updated during
the emitted state update. However when you use the `JUMP_TO` action to
go back to a previous screen history (important, as Tab keeps screens
rendered) inside a nested tab navigators, React Navigation performs the
logic in a different order. It first fires the state update and then
animates to the screen.

Because of this, the screen "updates" with the incorrect information,
and then the correct routeInfo is populated but the screen isn't
re-rendered.

Why this only occurs with the touch based navigation I'm not sure, I
would need to dig further into React Navigation to understand the
timings here.

This fix changed the RouterStore to calculate routeInfo in either the
state update (if the state isn't stale) or for the focused screen.

# Test Plan

Added unit test to cover both navigation via the imperative API and via
the touch api.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)