[router] Include tests in typecheck. Use seperate build tsconfig (#36485)
# Why

Since the SDK53 TypeScript upgrade, Expo Router's test files have been
showing in-editor only Type Errors. This was because the `__test__`
folder was excluded causing these files to use a default
`tsconfig.json`.

This is a VSCode / IDE error only - actual builds worked correctly.

This PR remove `__tests__` from the default `tsconfig.json` and uses a
new `tsconfig.build.json` for building the library (which excludes the
tests).

It also fixes some TypeScript errors in certain tests. There is still
one remaining error but this is a bug with router's types and will be
fixed separately, but because we only use `yarn build` to validate PRs
this bug won't be picked up.