[font] Bail out early if no loaded fonts (#36586)
# Why

Following up again from #36575 - it seems our mock for `getLoadedFonts`
will not return an empty array. @aleqsio can follow up with a fix and
then we can revert this. For now, I just handle this case in code and
bail out, this doesn't change any existing behavior of expo-font except
in a contrived test case scenario.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>