fix(metro-config): make process env readable in dev (#36652)
# Why

- fix https://github.com/expo/expo/issues/36503

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- Make process.env object readable in development. We avoided this
originally to prevent users from doing something in dev that isn't
possible in prod. We now have an eslint rule for this instead. We can
revisit making this stricter in the future.


<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

- `EXPO_ATLAS=1 EXPO_PUBLIC_TEST=my_value yarn start -w -c` Test that
`process.env.EXPO_PUBLIC_TEST` shows in development for server, web.

- `EXPO_PUBLIC_TEST=my_value npx expo export -p web -c` Test that
`process.env.EXPO_PUBLIC_TEST` shows in production for server, web.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>