[speech][Android] Fix calling getVoices before TTS initialization (#36664)
# Why

Calling the `getVoices` function before the TTS is initialized causes
`textToSpeech.voices` to throw an error. As a result, the function
returns an empty array instead of the list of available voices.

Resolves #36555.

# How

Although it's not the prettiest solution, we need to do a similar thing
to what we do in the `speak` function. We're adding promises to a queue,
and once TTS is initialized, all promises in the queue will be resolved.

# Test Plan

NCL -> TextToSpeech screen. The list of available voices should load
correctly.