[background-task] simplification of enqueuing tasks (#36705)
Tasks are only run in the background on iOS so we don't need to wait
with enqueuing workers until the app is backgrounded.

This commit fixes this by simplifying when app workers are enqueued:

- Enqueue worker for the first task registered
- Stop worker for the last task registered.
- Removed battery constraint on Android (this is false by default)
- Added rescheduling after 15-60 mins (depending on minimumInterval) if
app is foregrounded when task starts on Android

Closes ENG-15665

Test in BareExpo

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Łukasz Kosmaty <lukasz.kosmaty@swmansion.com>