[background-task] Expose testing method (#36732)
To be able to troubleshoot we need a way to trigger background tasks.

This commit exposes a testing method on both iOS and Android. The
function is called `triggerTaskWorkerForTestingAsync` and works by
triggering the BGTaskScheduler on iOS and the Expo TaskManager on
Android. The function was already implemented in the native code, but
not exposed to the user.

A documentation update is on its way as well.

Test in BareExpo

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>