[router] Fix web url from updating during initial load with nested navigators (#36690)
# Why

Fix https://x.com/ThetAungDev/status/1919442439758082404

# Problem

During initial render React Navigation performs multiple updates to the
URL. There are 3 locations for this all within `useLinking()` (the
locations are the code changes in this PR). All 3 have the same problem
but the code is different for each, so I'll just walk through one of
them.


https://github.com/react-navigation/react-navigation/blob/main/packages/native/src/useLinking.tsx#L345C1-L362C50

```
const onStateChange = async () => {
  const navigation = ref.current;

  if (!navigation || !enabled) {
    return;
  }

  const previousState = previousStateRef.current;

  /**

  This line here is the problem. Its getting the state from the rootState.

  During the initial render, the navigators navigate from the root downwards. They don't know the state of their children until the child has rendered. This is known as stale state.

  However, once they render they will fire this `onStateChange` call. Remember that their child state is currently undefined.
  */
  const state = navigation.getRootState();


  if (!state) {
    return;
  }

  const pendingPath = pendingPopStatePathRef.current;
  const route = findFocusedRoute(state);

  /**
    We have now passing the stale state into getPathForRoute
  */ 
  const path = getPathForRoute(route, state);
```


https://github.com/react-navigation/react-navigation/blob/09d65bcdc908883c8cdd1b93fde19cec4b799ad6/packages/native/src/useLinking.tsx#L306-L308
```
/**
  This is inside the getPathForRouteFunction.

   We are now asking the getPathFromState function to generate a path on the STALE state (the state within nested navigators).

   If this is a group, it will assume that we are going to the FIRST route within the group and return back an incorrect path.
*/
if (path == null) {
  path = getPathFromStateRef.current(state, configRef.current);
}
```

Since we have an incorrect path based upon stale state, the URL in
updated.

The process repeats for all nested navigators until it gets to the last
navigator

This navigator has the all the state, and therefore generates the
correct URL.


So from the users perspective,

- The URL is correct when they first load the page
- The url is updated many times to incorrect values (the flashes)
- The url is then correct


# Fix

I've updated the logic to use the store's state, which is the focused
state. The focused state is never stale and has the entire state tree.

Now each navigator will generate the correct path and the URL will be
"updated" with the same value (which has no effect).

# Test Plan

I'm struggling to test this in playwright. Playwright does not seem to
detect these URL changes.

I've manually tested with my own projects